Change sketching of read to match refseq sketch
I ran through this tutorial and I got a warning and an error about the K size of the reads (21) vs the K size of reqseq (16). I also got a warning about the sketch size. I've added a change to the docs which appears to match the precomputed refseq sketch.